+++ .gitignore
@@ -1,3 +1,3 @@
-node_modules/
-*.iml
-.idea
+.vscode
+dist
+node_modules
