--- docs/pull_request_template.md
@@ -1,5 +1,0 @@
-- Add a meaningful entry to the changelog in the README.md.
-- You can add yourself to the list of contributors in the package.json if you
-  like. If you don't do it, you won't be mentioned.
-
-Thanks!
