+++ package.json
@@ -1,57 +1,23 @@
 {
-  "name": "clone",
-  "description": "deep cloning of objects and arrays",
-  "tags": [
-    "clone",
-    "object",
-    "array",
-    "function",
-    "date"
-  ],
-  "version": "2.1.2",
+  "name": "@wings-j/clone",
+  "version": "0.0.0",
+  "type": "module",
+  "main": "dist/index.js",
+  "author": "WingsJ",
+  "description": "Deeply clone arbitrary objects in Javascript.",
+  "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "git://github.com/pvorb/node-clone.git"
+    "url": "git://github.com/wings-j/clone.git"
   },
-  "bugs": {
-    "url": "https://github.com/pvorb/node-clone/issues"
-  },
-  "main": "clone.js",
-  "author": "Paul Vorbach <paul@vorba.ch> (http://paul.vorba.ch/)",
-  "contributors": [
-    "Blake Miner <miner.blake@gmail.com> (http://www.blakeminer.com/)",
-    "Tian You <axqd001@gmail.com> (http://blog.axqd.net/)",
-    "George Stagas <gstagas@gmail.com> (http://stagas.com/)",
-    "Tobiasz Cudnik <tobiasz.cudnik@gmail.com> (https://github.com/TobiaszCudnik)",
-    "Pavel Lang <langpavel@phpskelet.org> (https://github.com/langpavel)",
-    "Dan MacTough (http://yabfog.com/)",
-    "w1nk (https://github.com/w1nk)",
-    "Hugh Kennedy (http://twitter.com/hughskennedy)",
-    "Dustin Diaz (http://dustindiaz.com)",
-    "Ilya Shaisultanov (https://github.com/diversario)",
-    "Nathan MacInnes <nathan@macinn.es> (http://macinn.es/)",
-    "Benjamin E. Coe <ben@npmjs.com> (https://twitter.com/benjamincoe)",
-    "Nathan Zadoks (https://github.com/nathan7)",
-    "Róbert Oroszi <robert+gh@oroszi.net> (https://github.com/oroce)",
-    "Aurélio A. Heckert (http://softwarelivre.org/aurium)",
-    "Guy Ellis (http://www.guyellisrocks.com/)",
-    "fscherwi (https://fscherwi.github.io)",
-    "rictic (https://github.com/rictic)",
-    "Martin Jurča (https://github.com/jurca)",
-    "Misery Lee <miserylee@foxmail.com> (https://github.com/miserylee)",
-    "Clemens Wolff (https://github.com/c-w)",
-    "Sabin Thomas (https://github.com/sabinthomas)"
-  ],
-  "license": "MIT",
-  "engines": {
-    "node": ">=0.8"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "nodeunit": "~0.9.0"
-  },
-  "optionalDependencies": {},
   "scripts": {
-    "test": "nodeunit test.js"
+    "dev": "tsc --watch",
+    "build": "tsc",
+    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules npx jest --watch"
+  },
+  "devDependencies": {
+    "cross-env": "^10.0.0",
+    "jest": "^30.0.5",
+    "typescript": "^5.9.2"
   }
 }
